Make this potentially dangerous script a little safer.
git-svn-id: https://svn.code.sf.net/p/python-ogre/code/trunk@834 503f2841-2733-4260-b5db-97e3c2726e36

+++ python-ogre/scripts/build-maketar.sh
@@ -1,5 +1,10 @@
 #! /bin/sh
 
+if [ ! -e python-ogre-pristine ]; then
+  echo "I need a clean checkout of python-ogre in python-ogre-pristine"
+  exit 1
+fi
+
 rm -rvf python-ogre
 
 cd python-ogre-pristine
